+++ anomalib/deploy/inferencers/torch_inferencer.py
@@ -13,10 +13,12 @@ from omegaconf import DictConfig, ListConfig
 from torch import Tensor
 
 from anomalib.config import get_configurable_parameters
+from anomalib.data import TaskType
+from anomalib.data.utils import InputNormalizationMethod, get_transforms
+from anomalib.data.utils.boxes import masks_to_boxes
 from anomalib.deploy.export import get_model_metadata
 from anomalib.models import get_model
 from anomalib.models.components import AnomalyModule
-from anomalib.pre_processing import PreProcessor
 
 from .base_inferencer import Inferencer
 
@@ -59,7 +61,8 @@ class TorchInferencer(Inferencer):
 
         self.meta_data = self._load_meta_data(meta_data_path)
 
-    def _get_device(self, device: str) -> torch.device:
+    @staticmethod
+    def _get_device(device: str) -> torch.device:
         """Get the device to use for inference.
 
         Args:
@@ -118,11 +121,18 @@ class TorchInferencer(Inferencer):
             Tensor: pre-processed image.
         """
         transform_config = (
-            self.config.dataset.transform_config.val if "transform_config" in self.config.dataset.keys() else None
+            self.config.dataset.transform_config.eval if "transform_config" in self.config.dataset.keys() else None
         )
-        image_size = tuple(self.config.dataset.image_size)
-        pre_processor = PreProcessor(transform_config, image_size)
-        processed_image = pre_processor(image=image)["image"]
+
+        image_size = (self.config.dataset.image_size[0], self.config.dataset.image_size[1])
+        center_crop = self.config.dataset.get("center_crop")
+        if center_crop is not None:
+            center_crop = tuple(center_crop)
+        normalization = InputNormalizationMethod(self.config.dataset.normalization)
+        transform = get_transforms(
+            config=transform_config, image_size=image_size, center_crop=center_crop, normalization=normalization
+        )
+        processed_image = transform(image=image)["image"]
 
         if len(processed_image) == 3:
             processed_image = processed_image.unsqueeze(0)
@@ -197,9 +207,18 @@ class TorchInferencer(Inferencer):
             if pred_mask is not None:
                 pred_mask = cv2.resize(pred_mask, (image_width, image_height))
 
+        if self.config.dataset.task == TaskType.DETECTION:
+            pred_boxes = masks_to_boxes(torch.from_numpy(pred_mask))[0][0].numpy()
+            box_labels = np.ones(pred_boxes.shape[0])
+        else:
+            pred_boxes = None
+            box_labels = None
+
         return {
             "anomaly_map": anomaly_map,
             "pred_label": pred_label,
             "pred_score": pred_score,
             "pred_mask": pred_mask,
+            "pred_boxes": pred_boxes,
+            "box_labels": box_labels,
         }