+++ anomalib/models/components/classification/__init__.py
@@ -1,0 +1,5 @@
+"""Classification modules."""
+
+from .kde_classifier import FeatureScalingMethod, KDEClassifier
+
+__all__ = ["KDEClassifier", "FeatureScalingMethod"]