+++ tests/pre_merge/models/test_model_premerge.py
@@ -17,6 +17,7 @@ class TestModel:
     @pytest.mark.parametrize(
         ["model_name", "nncf"],
         [
+            ("cfa", False),
             ("cflow", False),
             ("csflow", False),
             ("dfkde", False),
@@ -27,6 +28,7 @@ class TestModel:
             ("padim", False),
             ("patchcore", False),
             ("reverse_distillation", False),
+            ("rkde", False),
             ("stfpm", False),
         ],
     )