+++ tools/inference/torch_inference.py
@@ -47,7 +47,7 @@ def get_args() -> Namespace:
         required=False,
         help="Task type.",
         default="classification",
-        choices=["classification", "segmentation"],
+        choices=["classification", "detection", "segmentation"],
     )
     parser.add_argument(
         "--visualization_mode",